Fixed multithreading and simplify the inheritance scheme
The multithreading is now always going to return corretly even when
canceled with the stop button for the circle approximators

The reworking of the classes to move some of the basics into the
base classes is done. Also a little different flow to having the
approximators started. This reduced copied/pasted code and is
less likely to have the same thing s happen in the future.